+++ apps/angular/projection/src/app/component/city-card/city-card.component.ts
@@ -1,11 +1,9 @@
 import { AsyncPipe } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
-import { CityStore } from '../../data-access/city.store';
-import { randomCity } from '../../data-access/fake-http.service';
-import { City } from '../../model/city.model';
+import { CityStore, randomCity } from '../../data-access';
+import { City } from '../../model';
 import { DataCardComponentBase, DataStoreBase } from '../../shared';
-import { CardComponent } from '../../ui/card/card.component';
-import { ListItemComponent } from '../../ui/list-item/list-item.component';
+import { CardComponent, ListItemComponent } from '../../ui';
 
 @Component({
   selector: 'app-city-card',
