+++ apps/angular/projection/src/app/data-access/index.ts
@@ -1,0 +1,4 @@
+export * from './city.store';
+export * from './fake-http.service';
+export * from './student.store';
+export * from './teacher.store';
