--- apps/angular/projection/src/app/model/card.model.ts
@@ -1,13 +1,0 @@
-export enum CardType {
-  TEACHER,
-  STUDENT,
-  CITY,
-}
-
-export class CardItem {
-  constructor(
-    public readonly id: number,
-    public readonly name: string,
-    public readonly type?: CardType,
-  ) {}
-}
