+++ apps/angular/projection/src/app/model/index.ts
@@ -1,0 +1,3 @@
+export * from './city.model';
+export * from './student.model';
+export * from './teacher.model';
