+++ apps/angular/projection/src/app/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './card/card.component';
+export * from './list-item/list-item.component';
